Content Factory Pro/Code Files/Claude-Code-Configs/automation/requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
anthropic>=0.18.0
google-api-python-client>=2.70.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0